Add a warning on exit from chroot about proper cleanup
Another attempt to fix this correctly failed. Seems like it is related to some
kind of interaction between forking by unshare and job control. It is possible
to avoid this by using dash and disabling monitoring (+m option) but I've not
found a way to start Bash without the monitoring.